requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
